+++ android/KMEA/app/src/main/java/com/tavultesoft/kmea/KMManager.java
@@ -104,6 +104,7 @@ public final class KMManager {
   private static boolean debugMode = false;
   private static boolean shouldAllowSetKeyboard = true;
   private static boolean didCopyAssets = false;
+
   private static GlobeKeyAction inappKbGlobeKeyAction = GlobeKeyAction.GLOBE_KEY_ACTION_SHOW_MENU;
   private static GlobeKeyAction sysKbGlobeKeyAction = GlobeKeyAction.GLOBE_KEY_ACTION_SHOW_MENU;
   private static int sysKbIndexOnLockScreen = -1;
@@ -239,7 +240,7 @@ public final class KMManager {
 
     mFirebaseAnalytics = FirebaseAnalytics.getInstance(context);
 
-    if (!didCopyAssets) {
+    if (!didCopyAssets || isTestMode()) {
       copyAssets(appContext);
       migrateOldKeyboardFiles(appContext);
       updateOldKeyboardsList(appContext);
@@ -403,9 +404,11 @@ public final class KMManager {
   public static void onDestroy() {
     if (InAppKeyboard != null) {
       InAppKeyboard.onDestroy();
+      InAppKeyboard = null;
     }
     if (SystemKeyboard != null) {
       SystemKeyboard.onDestroy();
+      SystemKeyboard = null;
     }
     CloudDownloadMgr.getInstance().shutdown(appContext);
   }
@@ -956,7 +959,7 @@ public final class KMManager {
     boolean result1 = false;
     boolean result2 = false;
 
-    if (InAppKeyboard != null && InAppKeyboardLoaded)
+    if (InAppKeyboard != null && (InAppKeyboardLoaded || isTestMode()))
       result1 = InAppKeyboard.setKeyboard(packageID, keyboardID, languageID, keyboardName, languageName, kFont, kOskFont);
 
     if (SystemKeyboard != null && SystemKeyboardLoaded)
@@ -1205,6 +1208,13 @@ public final class KMManager {
 
   public static boolean isDebugMode() {
     return debugMode;
+  }
+
+  /**
+   * @return check for unit test.
+   */
+  public static boolean isTestMode() {
+    return Boolean.parseBoolean(System.getProperty("kmeaTestMode"));
   }
 
   public static void setShouldAllowSetKeyboard(boolean value) {