+++ android/history.md
@@ -14,6 +14,7 @@
   * Update target Android SDK version to 29 (#2279)
   * Add linting to Debug builds and resolve lint errors (#2305)
   * Fix memory issues during build process (#2361)
+  * Add simple UI tests for keyboard picker and keyboard info screens (#2326)
 
 ## 2019-10-30 12.0.4206 stable
 * Bug fix: